refactoring Checkout3DsComponent > Checkout3DsModal to support the hardware back in modal
